// src/types/movie.tsx
interface MovieRating {
	Source: string;
	Value: string;
}

interface Movie {
	Title: string;
	Year: string;
	Rated: string;
	Released: string;
	Runtime: string;
	Genre: string;
	Director: string;
	Writer: string;
	Actors: string;
	Plot: string;
	Language: string;
	Country: string;
	Awards: string;
	Poster: string;
	Ratings: MovieRating[];
	Metascore: string;
	imdbRating: string;
	imdbVotes: string;
	imdbID: string;
	Type: string;
	DVD: string;
	BoxOffice: string;
	Production: string;
	Website: string;
	Response: string;
}

interface MovieSearch {
	Search: Movie[];
	totalResults: string;
	Response: string;
}

export type { Movie, MovieSearch };